Move demangle into libmesh_common
